Extract batch tests to separate test file
