+++ tests/test_entity_operations.rs
@@ -129,129 +129,3 @@ async fn test_entity_operations() -> Result<()> {
 
     Ok(())
 }
-
-#[tokio::test]
-#[serial]
-async fn test_concurrent_entity_creation_batch() -> Result<()> {
-    init_logger(false);
-
-    // Start Arkiv container
-    let container = ArkivContainer::new(Config::default()).await?;
-    let client = ArkivClient::new(container.get_url()?)?;
-    let account = create_test_account(&client).await?;
-
-    // Number of entities to create per task
-    const ENTITIES_PER_TASK: usize = 15;
-
-    // Spawn two tasks that will create entities concurrently using create_entry
-    let task1 = tokio::spawn({
-        let client = client.clone();
-        let account = account;
-        async move {
-            let mut entity_ids = Vec::new();
-            for i in 0..ENTITIES_PER_TASK {
-                let payload = format!("task1_entity_{}", i).into_bytes();
-                let entry = Create::new(payload, 300)
-                    .annotate_string("task", "task1")
-                    .annotate_number("index", i as u64);
-                let entity_id = client.create_entry(account, entry).await?;
-                entity_ids.push(entity_id);
-            }
-            Ok::<_, anyhow::Error>(entity_ids)
-        }
-    });
-
-    let task2 = tokio::spawn({
-        let client = client.clone();
-        let account = account;
-        async move {
-            let mut entity_ids = Vec::new();
-            for i in 0..ENTITIES_PER_TASK {
-                let payload = format!("task2_entity_{}", i).into_bytes();
-                let entry = Create::new(payload, 300)
-                    .annotate_string("task", "task2")
-                    .annotate_number("index", i as u64);
-                let entity_id = client.create_entry(account, entry).await?;
-                entity_ids.push(entity_id);
-            }
-            Ok::<_, anyhow::Error>(entity_ids)
-        }
-    });
-
-    // Wait for both tasks to complete
-    let (task1_results, task2_results) = tokio::join!(task1, task2);
-    let task1_entity_ids = task1_results??;
-    let task2_entity_ids = task2_results??;
-
-    // Verify all entities were created successfully
-    for (i, entity_id) in task1_entity_ids.iter().enumerate() {
-        let entry_str = client.cat(*entity_id).await?;
-        assert_eq!(entry_str, format!("task1_entity_{}", i));
-
-        let metadata = client.get_entity_metadata(*entity_id).await?;
-        assert_eq!(metadata.string_annotations[0].value, "task1");
-        assert_eq!(metadata.numeric_annotations[0].value, i as u64);
-    }
-
-    for (i, entity_id) in task2_entity_ids.iter().enumerate() {
-        let entry_str = client.cat(*entity_id).await?;
-        assert_eq!(entry_str, format!("task2_entity_{}", i));
-
-        let metadata = client.get_entity_metadata(*entity_id).await?;
-        assert_eq!(metadata.string_annotations[0].value, "task2");
-        assert_eq!(metadata.numeric_annotations[0].value, i as u64);
-    }
-
-    log::info!(
-        "Successfully verified {} concurrent single entity creations",
-        ENTITIES_PER_TASK * 2
-    );
-    Ok(())
-}
-
-#[tokio::test]
-#[serial]
-async fn test_batch_entity_creation() -> Result<()> {
-    init_logger(false);
-
-    // Start Arkiv container
-    let container = ArkivContainer::new(Config::default()).await?;
-    let client = ArkivClient::new(container.get_url()?)?;
-    let account = create_test_account(&client).await?;
-
-    // Create multiple entities in a single batch
-    let mut creates = Vec::new();
-    for i in 0..5 {
-        let payload = format!("batch_entity_{}", i).into_bytes();
-        let entry = Create::new(payload, 300)
-            .annotate_string("batch", "test")
-            .annotate_number("index", i as u64);
-        creates.push(entry);
-    }
-
-    // Use the batch creation function that returns entity IDs
-    let entity_ids = client.create_entries(account, creates).await?;
-
-    // Verify we got the expected number of entity IDs
-    assert_eq!(entity_ids.len(), 5);
-
-    // Verify each entity ID and retrieve the data.
-    // This checks if entity IDs were retrieved in order.
-    for (i, entity_id) in entity_ids.iter().enumerate() {
-        // Verify we can retrieve the entity data
-        let entry_str = client.cat(*entity_id).await?;
-        assert_eq!(entry_str, format!("batch_entity_{}", i));
-
-        // Verify metadata
-        let metadata = client.get_entity_metadata(*entity_id).await?;
-        assert_eq!(metadata.string_annotations[0].value, "test");
-        assert_eq!(metadata.numeric_annotations[0].value, i as u64);
-        assert_eq!(metadata.owner.unwrap(), account);
-    }
-
-    log::info!(
-        "Successfully verified batch entity creation with {} entity IDs",
-        entity_ids.len()
-    );
-    Ok(())
-}
